requirements.txt
darkdetect==0.7.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyqtdarktheme==2.1.0
PyYAML==6.0
